adding info modal & add methods to homeController

+++ database/migrations/2022_05_14_094339_create_bills_table.php
@@ -18,7 +18,7 @@ class CreateBillsTable extends Migration
             $table->integer('user_id');
             $table->string('payment_date');
             $table->string('month_name');
-            $table->float('bill_cost');
+            $table->double('bill_cost');
             $table->boolean('payment_status')->default(false);
             $table->timestamps();
         });
